backend/account/pyproject.toml
[tool.poetry]
name = "account-api"
version = "0.1.0"
description = "account api for learning management system"
authors = ["a.espahbodi <dev.amir.espahbodi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
environs = {extras = ["django"], version = "^11.0.0"}
psycopg = {extras = ["binary"], version = "^3.1.19"}
django = "^5.0.6"
django-phonenumber-field = "^7.3.0"
django-phonenumbers = "^1.0.1"
user-agents = "^2.2.0"
gunicorn = "^22.0.0"
uvicorn = "^0.29.0"
orjson = "^3.10.3"
pydantic = {extras = ["email"], version = "^2.7.1"}
python-ipware = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
